requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
PyYAML>=6.0
python-dateutil>=2.8.0

# Data processing and analysis
openpyxl>=3.1.0  # For Excel export
xlsxwriter>=3.1.0  # For Excel formatting

# Utilities
pathlib2>=2.3.7; python_version < '3.4'
typing-extensions>=4.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For advanced features
# scikit-learn>=1.3.0  # For ML-based confidence calibration
# matplotlib>=3.7.0  # For visualization
# seaborn>=0.12.0  # For advanced plots